move tests to MappingTests.cs as we tend to move away from issuexxxx files
